@@ -3,7 +3,7 @@ package csc_cccix.geocracy.game;
 import android.content.Context;
 import android.opengl.GLES20;
 import android.opengl.GLES30;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.FragmentManager;
 import android.util.Log;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -114,6 +114,7 @@ public class Game implements Serializable {
 
     // GETTERS
     public GameActivity getActivity() { return activity; }
+    public GameStateMachine getStateMachine() { return StateMachine; }
     public CameraController getCameraController() { return cameraController; }
     public World getWorld() { return world; }
     public Player[] getPlayers() { return players; }
